% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Alpha_and_Beta.R
\name{alp_bet}
\alias{alp_bet}
\title{Alpha and Beta of Portfolio}
\usage{
alp_bet(ticker, wts, from, to, bench, free)
}
\arguments{
\item{wts}{This should be a double or a vector of doubles. It is the weights for each according ticker in your portfolio.}

\item{from}{This should be a string in form of 'yyyy-mm-dd'. It is the start of the time interval of your portfolio.}

\item{to}{This should be a string in form of 'yyyy-mm-dd'. It is the end of the time interval of your portfolio.}

\item{bench}{This should be a string. It is the ticker of the bench stock which is considered as the expected return.}

\item{free}{This should be a double. It is the risk-free interest rate.}

\item{tickers}{This should be a string or a vector of strings. It is the tickers of stocks you want to look at.}
}
\description{
Alpha and Beta of Portfolio
}
\examples{
alp_bet(c('JNJ' , 'WFC' , 'KO') , c(0.4 , 0.4 , -0.2) , '2020-11-11' , '2020-12-11' , 'SPX' , 0.09)
}
